Fix documentation for SNS/SQS scenario
This commit fixes a documentation issue for the SNS/SQS scenario: You
won't need the topic when starting the subscriber.

We also add the STS service in localstack. It's not required for now,
but it helps you introspecting your AWS account details, e.g. your AWS
Account ID (which you might need when building e.g. an ARN dynamically).